+++ BoxCLI/Commands/SharedLinkSubCommands/FilesSharedLinkSubCommandFactory.cs
@@ -14,7 +14,7 @@ namespace BoxCLI.Commands.SharedLinkSubCommands
 
         public override ISubCommand CreateSubCommand(string commandName)
         {
-            if (commandName == base._names.SubCommandNames.Get) 
+            if (commandName == base._names.SubCommandNames.Get)
             {
                 return new SharedLinkGetCommand(_boxPlatformBuilder, _boxHome, _names, BoxType.file);
             }
@@ -22,10 +22,14 @@ namespace BoxCLI.Commands.SharedLinkSubCommands
             {
                 return new SharedLinkCreateCommand(_boxPlatformBuilder, _boxHome, _names, BoxType.file);
             }
-			else if (commandName == base._names.SubCommandNames.Update)
-			{
-				return new SharedLinkUpdateCommand(_boxPlatformBuilder, _boxHome, _names, BoxType.file);
-			}
+            else if (commandName == base._names.SubCommandNames.Update)
+            {
+                return new SharedLinkUpdateCommand(_boxPlatformBuilder, _boxHome, _names, BoxType.file);
+            }
+            else if (commandName == base._names.SubCommandNames.Delete)
+            {
+                return new SharedLinkDeleteCommand(_boxPlatformBuilder, _boxHome, _names, BoxType.file);
+            }
             else
             {
                 throw new Exception("Command not registered.");