+++ BoxCLI/Commands/SharedLinkSubCommands/SharedLinkUpdateCommand.cs
@@ -12,6 +12,7 @@ namespace BoxCLI.Commands.SharedLinkSubCommands
     public class SharedLinkUpdateCommand : SharedLinkSubCommandBase
     {
         private CommandArgument _id;
+        private CommandArgument _type;
         private CommandOption _access;
         private CommandOption _password;
         private CommandOption _unsharedAt;
@@ -35,7 +36,10 @@ namespace BoxCLI.Commands.SharedLinkSubCommands
             _password = command.Option("--password <PASSWORD>", "Shared link password", CommandOptionType.SingleValue);
             _unsharedAt = command.Option("--unshared-at <TIME>", "Time that this link will become disabled, use formatting like 03w for 3 weeks.", CommandOptionType.SingleValue);
             _canDownload = command.Option("--can-download", "Whether the shared link allows downloads", CommandOptionType.NoValue);
-
+            if (base._t == BoxType.enterprise)
+            {
+                _type = command.Argument("itemType", "Type of item for shared link: either file or folder.");
+            }
             command.OnExecute(async () =>
             {
                 return await this.Execute();
@@ -52,10 +56,19 @@ namespace BoxCLI.Commands.SharedLinkSubCommands
         private async Task RunUpdate()
         {
             base.CheckForId(this._id.Value, this._app);
+            if (base._t == BoxType.enterprise)
+            {
+                if (this._type.Value == String.Empty && this._type.Value != SharedLinkSubCommandBase.BOX_FILE && this._type.Value != SharedLinkSubCommandBase.BOX_FOLDER)
+                {
+                    _app.ShowHelp();
+                    throw new Exception("You must provide an item type for this command: choose file or folder");
+                }
+            }
             var boxClient = base.ConfigureBoxClient(base._asUser.Value());
-            if (base._t == BoxType.file)
+            if (base._t == BoxType.file || (this._type != null && this._type.Value == SharedLinkSubCommandBase.BOX_FILE))
             {
                 var fileRequest = new BoxFileRequest();
+                fileRequest.Id = this._id.Value;
                 fileRequest.SharedLink = new BoxSharedLinkRequest();
                 if (this._access.HasValue())
                 {
@@ -81,11 +94,12 @@ namespace BoxCLI.Commands.SharedLinkSubCommands
                     return;
                 }
                 Reporter.WriteSuccess("Updated shared link:");
-                base.PrintItem(result);
+                base.PrintSharedLink(result.SharedLink);
             }
-            else if (base._t == BoxType.folder)
+            else if (base._t == BoxType.folder || (this._type != null && this._type.Value == SharedLinkSubCommandBase.BOX_FOLDER))
             {
                 var folderUpdateRequest = new BoxFolderRequest();
+                folderUpdateRequest.Id = this._id.Value;
                 folderUpdateRequest.SharedLink = new BoxSharedLinkRequest();
                 if (this._access.HasValue())
                 {
@@ -111,7 +125,7 @@ namespace BoxCLI.Commands.SharedLinkSubCommands
                     return;
                 }
                 Reporter.WriteSuccess("Updated shared link:");
-                base.PrintItem(updated);
+                base.PrintSharedLink(updated.SharedLink);
             }
             else
             {