+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.2
+- Added shared link commands for Create, Get, Update, and Delete. Located as standalone commands under `box shared-links` as well as under `box folders shared-links` and `box files shared-links`.
+- Fixed UTF-8 issue obscuring characters when saving CSV reports.
+- Added autopaging for events and fixed faulty event Get command.
+- Added date options `w` for weeks and `now` when using date options with commands.
+
 ## 1.0.1
 
 - Fixed minor bug preventing --name, --description, and --etag options from working on the `box folders update` command